+++ packages/hoop/src/factories/index.ts
@@ -4,3 +4,4 @@ export * from "./bindings.factory";
 export * from "./bits.factory";
 export * from "./dialog-elements.factory";
 export * from "./form-fields.factory";
+export * from "./hoop.factory";