+++ packages/hoop/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.1] - 2026-02-16
+
+### Fixed 
+- Stray console log in post builder.
+- Typos in post mutateAll fields method.
+- Hoop Factory export.
+
 ## [0.1.0] - 2026-02-14
 
 ### Added